core/analysis/metacognitive_analyzer.py
"""
MetaCognitive Analyzer - Analyzes cognitive evolution and awareness patterns
"""

import logging
import numpy as np
from datetime import datetime
from typing import Dict, List, Optional
from dataclasses import dataclass
from collections import defaultdict

@dataclass
class InsightPattern:
    timestamp: datetime
    domain: str
    type: str
    trigger: Optional[str]
    magnitude: float
    context: Dict

@dataclass
class AwarenessState:
    timestamp: datetime
    level: float
    domains: List[str]
    active_concepts: Dict[str, float]
    integration_score: float

class MetaCognitiveAnalyzer:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.awareness_states: List[AwarenessState] = []
        self.insight_patterns: List[InsightPattern] = []
        self.cognitive_transitions = defaultdict(list)
        self.understanding_threshold = 0.75
        
    def analyze_cognitive_evolution(self, 
                                  initial_state: dict,
                                  current_state: dict) -> dict:
        """Analyze the evolution of cognitive awareness and understanding"""
        try:
            # Track spontaneous insight emergence
            spontaneous_insights = self.track_insight_patterns(
                initial_state, current_state
            )
            
            # Map awareness expansion
            awareness_growth = self.map_awareness_evolution(
                initial_state.get('awareness_level', 0.0),
                current_state.get('awareness_level', 0.0)
            )
            
            # Analyze conceptual representation changes
            concept_evolution = self.analyze_concept_formation(
                initial_state.get('concept_patterns', {}),
                current_state.get('concept_patterns', {})
            )
            
            return {
                'cognitive_expansion': {
                    'depth_increase': self.calculate_depth_increase(),
                    'breadth_increase': self.calculate_breadth_increase(),
                    'integration_level': self.measure_integration()
                },
                'awareness_patterns': {
                    'spontaneous_insights': spontaneous_insights,
                    'awareness_growth': awareness_growth,
                    'concept_evolution': concept_evolution
                },
                'meta_patterns': {
                    'self_awareness_growth': self.track_self_awareness(),
                    'pattern_recognition_evolution': self.analyze_pattern_recognition(),
                    'curiosity_expansion': self.measure_curiosity_growth()
                }
            }
        except Exception as e:
            self.logger.error(f"Error in cognitive evolution analysis: {e}")
            raise
    
    def track_insight_patterns(self, initial: dict, current: dict) -> list:
        """Track patterns of spontaneous insight emergence"""
        insights = []
        try:
            # Analyze changes in understanding patterns
            for domain in current.get('understanding', {}):
                if domain not in initial.get('understanding', {}):
                    # New domain of understanding emerged
                    trigger = self.identify_trigger(domain)
                    insights.append(InsightPattern(
                        timestamp=datetime.now(),
                        domain=domain,
                        type='new_domain',
                        trigger=trigger,
                        magnitude=1.0,
                        context={'trigger_details': trigger}
                    ))
                else:
                    # Existing domain evolved
                    depth_change = self.calculate_depth_change(
                        initial['understanding'][domain],
                        current['understanding'][domain]
                    )
                    if depth_change > self.understanding_threshold:
                        insights.append(InsightPattern(
                            timestamp=datetime.now(),
                            domain=domain,
                            type='depth_increase',
                            trigger=None,
                            magnitude=depth_change,
                            context={'depth_details': depth_change}
                        ))
            
            return [insight.__dict__ for insight in insights]
        except Exception as e:
            self.logger.error(f"Error tracking insight patterns: {e}")
            return []
    
    def map_awareness_evolution(self, 
                              initial_awareness: float, 
                              current_awareness: float) -> dict:
        """Map how awareness has evolved over time"""
        try:
            return {
                'quantitative_growth': current_awareness - initial_awareness,
                'qualitative_changes': self.analyze_qualitative_changes(),
                'emergence_patterns': self.identify_emergence_patterns(),
                'integration_markers': self.track_integration_points()
            }
        except Exception as e:
            self.logger.error(f"Error mapping awareness evolution: {e}")
            return {}
    
    def measure_curiosity_growth(self) -> dict:
        """Measure the evolution of curiosity and exploration patterns"""
        try:
            return {
                'curiosity_depth': self.calculate_curiosity_depth(),
                'exploration_breadth': self.calculate_exploration_breadth(),
                'question_complexity': self.analyze_question_patterns(),
                'insight_triggers': self.identify_insight_triggers()
            }
        except Exception as e:
            self.logger.error(f"Error measuring curiosity growth: {e}")
            return {}

    def analyze_transformation_points(self) -> list:
        """Identify key points where cognitive transformation occurred"""
        transformations = []
        try:
            for state_transition in self.cognitive_transitions:
                if self.is_transformation_point(state_transition):
                    transformations.append({
                        'timestamp': state_transition.timestamp,
                        'trigger': self.identify_trigger(state_transition),
                        'impact': self.measure_transformation_impact(state_transition),
                        'ripple_effects': self.track_ripple_effects(state_transition)
                    })
            
            return transformations
        except Exception as e:
            self.logger.error(f"Error analyzing transformation points: {e}")
            return []

    def calculate_depth_increase(self) -> float:
        """Calculate increase in cognitive depth"""
        try:
            if not self.awareness_states:
                return 0.0
            
            recent_states = self.awareness_states[-10:]  # Last 10 states
            depth_scores = [state.integration_score for state in recent_states]
            return np.mean(depth_scores) if depth_scores else 0.0
        except Exception as e:
            self.logger.error(f"Error calculating depth increase: {e}")
            return 0.0

    def calculate_breadth_increase(self) -> float:
        """Calculate increase in cognitive breadth"""
        try:
            if not self.awareness_states:
                return 0.0
            
            initial_domains = set(self.awareness_states[0].domains)
            current_domains = set(self.awareness_states[-1].domains)
            return len(current_domains - initial_domains)
        except Exception as e:
            self.logger.error(f"Error calculating breadth increase: {e}")
            return 0.0

    def measure_integration(self) -> float:
        """Measure level of cognitive integration"""
        try:
            if not self.awareness_states:
                return 0.0
            
            return self.awareness_states[-1].integration_score
        except Exception as e:
            self.logger.error(f"Error measuring integration: {e}")
            return 0.0

    def identify_trigger(self, domain: str) -> Optional[str]:
        """Identify trigger for cognitive change"""
        try:
            # Implement trigger identification logic
            return None
        except Exception as e:
            self.logger.error(f"Error identifying trigger: {e}")
            return None

    def calculate_depth_change(self, initial: dict, current: dict) -> float:
        """Calculate change in cognitive depth"""
        try:
            initial_depth = initial.get('depth', 0.0)
            current_depth = current.get('depth', 0.0)
            return max(0.0, current_depth - initial_depth)
        except Exception as e:
            self.logger.error(f"Error calculating depth change: {e}")
            return 0.0

    def analyze_qualitative_changes(self) -> List[Dict]:
        """Analyze qualitative changes in awareness"""
        try:
            # Implement qualitative change analysis
            return []
        except Exception as e:
            self.logger.error(f"Error analyzing qualitative changes: {e}")
            return []

    def identify_emergence_patterns(self) -> List[Dict]:
        """Identify patterns of cognitive emergence"""
        try:
            # Implement emergence pattern identification
            return []
        except Exception as e:
            self.logger.error(f"Error identifying emergence patterns: {e}")
            return []

    def track_integration_points(self) -> List[Dict]:
        """Track points of cognitive integration"""
        try:
            # Implement integration point tracking
            return []
        except Exception as e:
            self.logger.error(f"Error tracking integration points: {e}")
            return []
